+++ Cargo.toml
@@ -22,5 +22,5 @@ imgui-winit-support = "0.1.0"
 
 
 
-[dependencies.sdl2-sys]
-features = ["use_mac_framework"]
+# [dependencies.sdl2-sys]
+# features = ["use_mac_framework"]
