+++ src/make_world.rs
@@ -1,8 +1,12 @@
+#![allow(unused_imports)]
+#![allow(unused_variables)]
+
 use crate::camera::PerspectiveCamera;
 use crate::lightsource::PointLight;
 use crate::material::{MatTex, Material};
 use crate::parser;
-use crate::shape::{Intersectable, Plane, Sphere, Triangle, Vec2, Vec3, Vertex};
+use crate::shape::{Intersectable, Plane, Sphere, Triangle};
+use crate::algebra::{Vec3, Vec2, Vertex};
 use crate::skybox::Skybox;
 use crate::texture_map;
 use crate::thruster;
@@ -11,28 +15,31 @@ pub fn make_world<'a>() -> Result<thruster::Thruster<'a>, String> {
 	let mut texture_map = texture_map::TextureMap::new();
 
 	let checker_handle = texture_map.load_image_from_file("./textures/checker.png")?;
+	let wood_handle = texture_map.load_image_from_file("./textures/wood.png")?;
 	let earth_handle = texture_map.load_image_from_file("./textures/earth.png")?;
 
-	// let skybox = Skybox::new([
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_bk.png")?,
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_ft.png")?,
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_up.png")?,
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_dn.png")?,
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_rt.png")?,
-	// 	texture_map.load_image_from_file("./skybox/miramar/miramar_lf.png")?,
-	// ]);
-	let skybox = Skybox::new([
-		texture_map.load_image_from_file("./skybox/Yokohama3/negx.png")?,
-		texture_map.load_image_from_file("./skybox/Yokohama3/posx.png")?,
-		texture_map.load_image_from_file("./skybox/Yokohama3/posy.png")?,
-		texture_map.load_image_from_file("./skybox/Yokohama3/negy.png")?,
-		texture_map.load_image_from_file("./skybox/Yokohama3/posz.png")?,
-		texture_map.load_image_from_file("./skybox/Yokohama3/negz.png")?,
-	]);
+     let skybox = Skybox::new([
+         texture_map.load_image_from_file("./skybox/miramar/miramar_bk.png")?,
+         texture_map.load_image_from_file("./skybox/miramar/miramar_ft.png")?,
+         texture_map.load_image_from_file("./skybox/miramar/miramar_up.png")?,
+         texture_map.load_image_from_file("./skybox/miramar/miramar_dn.png")?,
+         texture_map.load_image_from_file("./skybox/miramar/miramar_rt.png")?,
+         texture_map.load_image_from_file("./skybox/miramar/miramar_lf.png")?,
+     ]);
+	//let skybox = Skybox::new([
+		//texture_map.load_image_from_file("./skybox/Yokohama3/negx.png")?,
+		//texture_map.load_image_from_file("./skybox/Yokohama3/posx.png")?,
+		//texture_map.load_image_from_file("./skybox/Yokohama3/posy.png")?,
+		//texture_map.load_image_from_file("./skybox/Yokohama3/negy.png")?,
+		//texture_map.load_image_from_file("./skybox/Yokohama3/posz.png")?,
+		//texture_map.load_image_from_file("./skybox/Yokohama3/negz.png")?,
+	//]);
 
 	let checker_mattex = MatTex::from_handle(checker_handle, Vec2::new(1000.0, 1000.0));
+	let wood_mattex = MatTex::from_handle(wood_handle, Vec2::new(1000.0, 1000.0));
 	let earth_mattex = MatTex::from_handle(earth_handle, Vec2::new(1.0, 1.0));
 	let plane_mat = Material::reflective(checker_mattex);
+	let wood_mat = Material::diffuse(wood_mattex);
 	let earth_mat = Material::reflective(earth_mattex);
 
 	let red = MatTex::Color(Vec3::new(255.0, 0.0, 0.0));
@@ -57,21 +64,21 @@ pub fn make_world<'a>() -> Result<thruster::Thruster<'a>, String> {
 		texture: green,
 	};
 
-	let obj = parser::parse("./objs/ammo.obj".to_string());
+	let obj = parser::parse("./objs/sphere.obj".to_string());
 	let mut scene: Vec<Box<dyn Intersectable + Sync>> = Vec::new();
-	// for (avt, bvt, cvt) in obj.triangles.iter() {
-	// 	scene.push(Box::new(Triangle {
-	// 		a: Vertex::from_parsed(avt),
-	// 		b: Vertex::from_parsed(bvt),
-	// 		c: Vertex::from_parsed(cvt),
-	// 		material: red_mat,
-	// 	}))
-	// }
+     //for (avt, bvt, cvt) in obj.triangles.iter() {
+         //scene.push(Box::new(Triangle {
+             //a: Vertex::from_parsed(avt),
+             //b: Vertex::from_parsed(bvt),
+             //c: Vertex::from_parsed(cvt),
+             //material: red_mat,
+         //}))
+     //}
 	scene.extend::<Vec<Box<dyn Intersectable + Sync>>>(vec![
 		Box::new(Plane {
 			origin: Vec3::new(0.0, 0.0, 100.0),
 			normal: Vec3::new(0.0, 1.0, -1.0).normalized(),
-			material: plane_mat,
+			material: wood_mat,
 		}),
 		// Box::new(Plane {
 		// 	origin: Vec3::new(0.0, 100.0, 100.0),
@@ -91,6 +98,11 @@ pub fn make_world<'a>() -> Result<thruster::Thruster<'a>, String> {
 		Box::new(Sphere {
 			origin: Vec3::new(-75.0, 100.0, 50.0),
 			radius: 50.0,
+			material: red_mat,
+		}),
+		Box::new(Sphere {
+			origin: Vec3::new(-75.0, 300.0, 50.0),
+			radius: 100.0,
 			material: red_mat,
 		}),
 		// Box::new(Sphere {
